# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "IFMeetup API",
      "description": "API para gerenciamento de eventos do IFPB",
      "contact": {
        "name": "IFPB",
        "url": "https://www.ifpb.edu.br",
        "email": "contato@ifpb.edu.br"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "https://www.apache.org/licenses/LICENSE-2.0"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Generated server url"
      }
    ],
    "security": [
      {
        "JWT": []
      }
    ],
...
    "paths": {
...
      "/auth/reset-password": {
        "post": {
          "tags": [
            "Autenticação"
          ],
          "summary": "Redefinição de senha",
          "description": "Redefine a senha de um usuário usando um token de recuperação",
          "operationId": "resetPassword",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Senha redefinida com sucesso",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Token inválido ou expirado",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/register": {
        "post": {
          "tags": [
            "Autenticação"
          ],
          "summary": "Registro de usuário",
          "description": "Registra um novo usuário no sistema",
          "operationId": "register",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Usuário registrado com sucesso",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Dados inválidos ou email já cadastrado",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      },
...
      "/auth/login": {
        "post": {
          "tags": [
            "Autenticação"
          ],
          "summary": "Login de usuário",
          "description": "Autentica um usuário no sistema",
          "operationId": "login",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "403": {
              "description": "Email não verificado",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Credenciais inválidas",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "200": {
              "description": "Login realizado com sucesso",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/forgot-password": {
        "post": {
          "tags": [
            "Autenticação"
          ],
          "summary": "Recuperação de senha",
          "description": "Envia um email com instruções para recuperação de senha",
          "operationId": "forgotPassword",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "422": {
              "description": "Email não encontrado",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "200": {
              "description": "Email enviado com sucesso",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      },
...
      "/auth/verify": {
        "get": {
          "tags": [
            "Autenticação"
          ],
          "summary": "Verificação de conta",
          "description": "Verifica uma conta de usuário usando um token enviado por email",
          "operationId": "verifyAccount",
          "parameters": [
            {
              "name": "token",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Conta verificada com sucesso",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Token inválido ou expirado",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/me": {
        "get": {
          "tags": [
            "Autenticação"
          ],
          "summary": "Obter usuário atual",
          "description": "Retorna os dados do usuário atualmente autenticado",
          "operationId": "me",
          "responses": {
            "200": {
              "description": "Dados do usuário recuperados com sucesso",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Usuário não autenticado",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "PasswordResetRequest": {
          "required": [
            "newPassword",
            "token"
          ],
          "type": "object",
          "properties": {
            "token": {
              "type": "string",
              "description": "Token de redefinição recebido por email",
              "example": "a1b2c3d4-5678-90ab-cdef-ghijklmnopqr"
            },
            "newPassword": {
              "maxLength": 2147483647,
              "minLength": 8,
              "type": "string",
              "description": "Nova senha (mínimo 8 caracteres)",
              "writeOnly": true,
              "example": "novaSenhaSecreta123"
            }
          },
          "description": "Dados para redefinição de senha"
        },
        "AuthResponse": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "description": "Mensagem informativa sobre o resultado da operação"
            },
            "user": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "sessionId": {
              "type": "string",
              "description": "Identificador único da sessão, quando aplicável",
              "format": "uuid"
            },
            "expiresAt": {
              "type": "string",
              "description": "Data e hora de expiração da sessão, quando aplicável",
              "format": "date-time"
            },
            "token": {
              "type": "string",
              "description": "Token JWT de autenticação, quando aplicável"
            }
          },
          "description": "Resposta padrão para operações de autenticação"
        },
        "UserResponse": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "Identificador único do usuário",
              "format": "uuid"
            },
            "name": {
              "type": "string",
              "description": "Nome completo do usuário"
            },
            "email": {
              "type": "string",
              "description": "Email do usuário"
            },
            "phoneNumber": {
              "type": "string",
              "description": "Número de telefone do usuário"
            },
            "roles": {
              "uniqueItems": true,
              "type": "array",
              "description": "Lista de papéis/roles do usuário no sistema",
              "items": {
                "type": "string",
                "description": "Lista de papéis/roles do usuário no sistema"
              }
            }
          },
          "description": "Resposta com dados básicos do usuário"
        },
        "RegisterRequest": {
          "required": [
            "email",
            "firstName",
            "lastName",
            "password",
            "profileType"
          ],
          "type": "object",
          "properties": {
            "firstName": {
              "maxLength": 50,
              "minLength": 2,
              "type": "string",
              "description": "Nome do usuário",
              "example": "Maria"
            },
            "lastName": {
              "maxLength": 50,
              "minLength": 2,
              "type": "string",
              "description": "Sobrenome do usuário",
              "example": "Silva"
            },
            "email": {
              "type": "string",
              "description": "Email do usuário",
              "example": "maria.silva@exemplo.com"
            },
            "phoneNumber": {
              "pattern": "^\\(?\\d{2}\\)?[\\s-]?\\d{4,5}-?\\d{4}$",
              "type": "string",
              "description": "Número de telefone do usuário",
              "example": "(83) 99999-9999"
            },
            "password": {
              "maxLength": 2147483647,
              "minLength": 8,
              "type": "string",
              "description": "Senha do usuário (mínimo 8 caracteres)",
              "writeOnly": true,
              "example": "senhaSecreta123"
            },
            "profileType": {
              "type": "string",
              "description": "Tipo de perfil do usuário",
              "example": "STUDENT",
              "enum": [
                "ADMIN",
                "COORDINATOR",
                "TEACHER",
                "STUDENT"
              ]
            }
          },
          "description": "Dados para registro de novo usuário"
        },
        "LoginRequest": {
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "description": "Email do usuário",
              "example": "usuario@exemplo.com"
            },
            "password": {
              "type": "string",
              "description": "Senha do usuário",
              "writeOnly": true,
              "example": "senhaSecreta123"
            }
          },
          "description": "Dados para autenticação de usuário"
        },
        "ForgotPasswordRequest": {
          "required": [
            "email"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            }
          }
        },
...
      },
      "securitySchemes": {
        "JWT": {
          "type": "http",
          "description": "JWT token para autenticação",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }